all locators and associated methods moved from MainPage to TopMenuPage
